Restructure condition dashboard files.
