+++ app/models/condition_dashboard/visibility.rb
@@ -1,3 +1,6 @@
 module Visibility
 
 end
+
+
+# Breakout of average number of rides, highest number of rides, and lowest number of rides on days with mean visibility in miles in four mile increments.
